requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
httpx==0.27.2
pydantic==2.10.2
pydantic-settings==2.6.1
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-instrumentation-logging==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-instrumentation-httpx==0.49b2
python-json-logger==2.0.7

# Test dependencies
pytest==7.4.4
pytest-asyncio==0.21.2
pytest-cov==4.1.0
